+++ packages/cedar/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.1.2
+
+- Add templates to `CedarPolicySet.toJson`
+
+## 0.1.1
+
+- Add template-linked policies
+
 ## 0.1.0
 
 - Initial version.